perf + feat: side-table date denormalization, trigram warmup, cursor fanout
### Perf (#64): popular-maintainer eq() queries

`eq('signed_off_by', 'kuba@kernel.org')` on netdev scale (28k+ SOB
rows) was taking 50-60 ms. EXPLAIN showed the plan walked every
matching row then did a TEMP B-TREE sort for the ORDER BY.

Fix:
  * Denormalize `date_unix_ns` into `over_trailer_email` and
    `over_touched_file`. Idempotent `ALTER TABLE ADD COLUMN` on
    reopen; write path populates on every INSERT.
  * New partial indexes `over_trailer_email_kind_email_date` and
    `over_touched_file_path_date` (both `WHERE date_unix_ns IS
    NOT NULL`) for the covering walk.
  * Rewrite both `scan_eq_via_*` paths as a `WITH picked AS (...)`
    CTE: inner subquery streams the top-N directly off the
    covering index, outer query does a bounded JOIN back to
    `over` for full MessageRows. No sort step.
  * `OverDb::backfill_side_table_dates()` + the
    `_core.backfill_side_table_dates(data_dir)` PyO3 binding copy
    dates from `over` into the side tables for existing rows.
    Chunked rowid-cursor walk with per-chunk HashMap lookups —
    the obvious single correlated-UPDATE approach generated a
    10+ GB WAL and never completed on lore scale.

### Perf (#70): trigram segment cache warmup at boot

`server.py::_warmup_tiers` already paged in BM25 and over.db.
Added a `reader.patch_search("__function__", None, 1)` probe so
the OS page cache holds the ~530 trigram segment files before
the first real request. Complements the per-process segment
cache shipped in #66; first `patch_search` no longer pays the
9s cold-mmap tail.

### Feat (#71, partial): cursor fanout to lore_patch_search + lore_regex

Both tools now accept the `cursor` parameter and emit
`next_cursor` when more pages remain. Same pattern as `lore_
search` (#67): query-scoped `query_hash`, 2× oversample,
newest-first date + mid tiebreak, mint from last returned row.

`CursorPayload.last_seen_score` widened from `f32` to `f64` so
nanosecond dates round-trip exactly. Backward-incompatible for
cursor tokens (we issued them for one session), but the HMAC
regime means tokens are expected to be ephemeral anyway.

`RowsResponse` now carries an optional `next_cursor` field so
the primitives that paginate (`lore_regex` for now) can emit
cursors through the same envelope as `SearchResponse`.

Follow-ups queued: wire cursors into `lore_activity`,
`lore_author_profile`, `lore_author_footprint` — the pattern is
established, remaining work is mechanical.

Tests: 145 Rust (unchanged) + 192 Python (unchanged), all green.

+++ src/router.rs
@@ -428,10 +428,18 @@ fn rrf_merge(tiers: HashMap<&'static str, Vec<MessageRow>>, limit: usize) -> Vec
 // --------------------------------------------------------------------
 // Cursor: HMAC-signed opaque pagination token.
 
+/// Pagination cursor payload.
+///
+/// `last_seen_score` is overloaded to carry either a relevance
+/// score (`lore_search` RRF) or a `date_unix_ns` tiebreak (the
+/// newest-first tools like `lore_patch_search`, `lore_activity`,
+/// `lore_regex`, `lore_author_footprint`). `f64` precision keeps
+/// nanosecond dates exact; for relevance scores in `[0, 1]` it's
+/// overkill but harmless.
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub struct CursorPayload {
     pub query_hash: u64,
-    pub last_seen_score: f32,
+    pub last_seen_score: f64,
     pub last_seen_mid: String,
 }
